requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioitertools==0.12.0
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.14.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
arize-phoenix==5.8.0
arize-phoenix-evals==0.17.3
arize-phoenix-otel==0.6.1
art==6.3
attrs==24.2.0
Authlib==1.3.2
bert-score==0.3.13
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.1.0
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.4
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.65.0
graphql-core==3.2.5
greenlet==3.1.1
grpc-interceptor==0.15.4
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-openai==0.2.6
langchain-text-splitters==0.3.2
langsmith==0.1.142
lxml==5.3.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.13.3
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.54.3
openinference-instrumentation==0.1.18
openinference-semantic-conventions==0.1.12
openpyxl==3.1.5
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-exporter-otlp-proto-http==1.28.1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==2.10.1
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.0.0
pycm==4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
ragas==0.1.9
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.21.0
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.4.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlean.py==3.47.0
starlette==0.41.2
strawberry-graphql==0.243.1
streamlit==1.40.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchdog==6.0.0
websockets==14.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.1
zipp==3.21.0
